pyproject.toml
[project]
name = "task-tickets"
version = "0.1.0"
description = ""
authors = [
    {name = "Evgenii Denisov",email = "eIGato@github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.11.0,<26.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]

[tool.poetry]
name = "task-tickets"
version = "0.1.0"
description = "Simple task management API"
authors = ["Evgenii Denisov <eIGato@github.com>"]
readme = "README.md"
packages = [{ include = "config" }, { include = "tasks" }]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = ">=5.2.8,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
djangorestframework-simplejwt = ">=5.5.1,<6.0.0"
drf-spectacular = ">=0.29.0,<0.30.0"
psycopg2-binary = ">=2.9.11,<3.0.0"

[tool.poetry.group.dev.dependencies]
black = ">=25.11.0,<26.0.0"
pytest = ">=9.0.1,<10.0.0"
pytest-django = ">=4.11.1,<5.0.0"

[tool.black]
line-length = 79
target-version = ["py313"]
exclude = """
/(
  .venv
  | migrations
  | __pycache__
)/
"""

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-ra --strict-markers"
